Modify some method implementation for vector (#8545)
Approved by: @ouyuanning, @m-schen, @XuPeng-SH, @nnsgmsone

+++ pkg/sql/plan/function/builtin/unary/values.go
@@ -23,9 +23,9 @@ func Values(parameters []*vector.Vector, result vector.FunctionResultWrapper, pr
 	fromVec := parameters[0]
 	toVec := result.GetResultVector()
 
-	sels := make([]int64, fromVec.Length())
+	sels := make([]int32, fromVec.Length())
 	for j := 0; j < len(sels); j++ {
-		sels[j] = int64(j)
+		sels[j] = int32(j)
 	}
 	toVec.Union(fromVec, sels, proc.GetMPool())
 	return nil
